Move Collector to Cloud::Collector
