+++ lib/topological_inventory/ansible_tower/logging.rb
@@ -8,8 +8,6 @@ module TopologicalInventory
 
     def self.logger
       @logger ||= TopologicalInventory::Providers::Common::Logger.new
-      @logger.level = Logger::DEBUG
-      @logger
     end
 
     module Logging
